Add prefix support
Use path of the url to make the prefix. This introduced an issue when
using the file backend ie file:///path/to/file. Here you'd want
/path/to/file to be set as the location. However it makes more sense to
use this as the prefix in the other backends. For now not using prefix
with the file backend until I can come up with something better.